+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- **EXPERIMENTAL_Modal** `ref` not properly passed to modal element.
+
 ## [9.133.1] - 2020-10-29
 
 ### Fixed