+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.133.2] - 2020-11-12
+
 ### Fixed
 
 - **EXPERIMENTAL_Modal** `ref` not properly passed to modal element.